Moved request handlers to specific package.
